requirements.txt
aiogram
aiohttp
python-dotenv
ollama
aiohttp-socks